% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_ancillary.R
\name{pdb2uniprot}
\alias{pdb2uniprot}
\title{Return the UniProt ID Given the PDB and Chain IDs}
\usage{
pdb2uniprot(pdb, chain)
}
\arguments{
\item{pdb}{the 4-letter PDB identifier.}

\item{chain}{letter identifying the chain.}
}
\value{
The function returns the UniProt ID for the chain of interest.
}
\description{
Returns the uniprot id of a given chain within a PDB structure.
}
\examples{
pdb2uniprot('1u8f', 'O')
}
\seealso{
uniprot2pdb(), id.mapping()
}
\author{
Juan Carlos Aledo
}
